Simplify rated property deserialization process.

+++ shared/src/commonMain/kotlin/com/example/moveeapp_compose_kmm/domain/account/favorite/MovieAccountState.kt
@@ -1,3 +1,3 @@
 package com.example.moveeapp_compose_kmm.domain.account.favorite
 
-data class MovieAccountState(val isFavorite: Boolean, val rating: Double?)
+data class MovieAccountState(val isFavorite: Boolean, val rating: Float?)
